pyproject.toml
[tool.poetry]
name = "mistral-pdf-to-markdown"
version = "0.1.0"
description = "A CLI tool to convert PDF files to Markdown using the Mistral AI OCR API."
authors = ["Arcangelo Massari <info@arcangelomassari.com>"]
readme = "README.md"
packages = [{include = "mistral_pdf_to_markdown", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
mistralai = "^1.7.0"
click = "^8.1.8"    
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pdf2md = "mistral_pdf_to_markdown.main:cli" 